Reduce decoder memory using multi-generation transfers
This implementaiton focuses on devices which don't have any flash to store
the (un)coded symbols and have a very limited amount of memory. Using a
single generation for all symbols is beneficial for the reliability
(efficiency) of the recovery process (of lost symbols) but needs a lot more
CPU time and (much worse for us) memory.

Instead of only having one generation of uncoded and coded symbols, it
splits up the sequence number range in multiple generations. One part of
such a sequence sub-range is just for uncoded packets and the rest of the
sub-range is for coded packets. If not the whole sub-range is used (by
transmitting coded packets), the missing sequence numbers are just skipped
during the transmission. For the decoder, this looks like lost coded
symbols - which are hopefully not needed for the recovery process.

Signed-off-by: Sven Eckelmann <sven@narfation.org>