Implement symbol number encoding limitation
If it is known already how many symbols need to be encoded, it is possible
to inform the encoder about this limitation. The encoder will then stop to
create new generations when it is not necessary to handle more uncoded
symbols. And the encoder will stop acccept more symbols for the current
generation when this limit was already reached.

Signed-off-by: Sven Eckelmann <sven@narfation.org>