Skip padding uncoded symbols
If the number of uncoded symbols the encoder has to encode is limited, then
we must now only encode 0 symbols for these padding symbols in the encoder.
But sending these are not useful when the receiver also knows the number of
uncoded symbols which need to be transmitted. The decoder can then simply
prefill these padding symbols with 0-symbols.

Signed-off-by: Sven Eckelmann <sven@narfation.org>