Implement symbol number decoding limitation
If it is known already how many symbols need to be decoded, it is possible
to inform the decoder about this limitation. The decoder will then stop to
create new generations when it is not necessary to handle more uncoded
symbols. And the decoder will stop searching for more decodable symbols.

Signed-off-by: Sven Eckelmann <sven@narfation.org>