Prefill decoder padding symbols
If the number of symbols the decoder has to decode is limited, then we must
now only encode 0 symbols for these padding symbols in the encoder. This
allows us to prefill these extra symbols without actually receiving them.
This can increase the chance that the last generation can be fully decoded
even when some symbols are not actually received.

Signed-off-by: Sven Eckelmann <sven@narfation.org>